refactor(pushpair): Move fetch last update to services
