+++ src/modules/pushPair/services/utils.ts
@@ -1,4 +1,5 @@
 import toPath from 'lodash.topath';
+import { utils } from "ethers";
 
 export function convertOldSourcePath(sourcePath: string): string {
     // Keep support for more functions
@@ -21,4 +22,9 @@ export function convertOldSourcePath(sourcePath: string): string {
     });
 
     return result;
+}
+
+export function computeFactoryPairId(pair: string, decimals: number) {
+    // Id = keccak256(bytes("Price-<PAIR>-<DECIMALS>)")
+    return utils.keccak256(utils.toUtf8Bytes("Price-" + pair + "-" + decimals.toString()))
 }
